requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
captum=0.2.0=0
cudatoolkit=10.2.89=hfd86e86_1
lime=0.2.0.0=pypi_0
matplotlib=3.2.1=pypi_0
matplotlib-base=3.1.3=py38hef1b27d_0
nltk=3.5=pypi_0
numpy=1.18.1=py38h4f9e942_0
numpy-base=1.18.1=py38hde5b4d6_1
python-papi=5.5.1.2=pypi_0
scikit-image=0.16.2=pypi_0
scikit-learn=0.22.2.post1=pypi_0
scipy=1.4.1=pypi_0
sentencepiece=0.1.86=pypi_0
tokenizers=0.5.2=pypi_0
torchvision=0.6.0=py38_cu102
tqdm=4.45.0=pypi_0
transformers=2.8.0=pypi_0